requirements.txt
asyncua>=1.0.0
psycopg2-binary
langchain>=0.2.0,<0.4.0
langchain-core>=0.2.0,<0.4.0
langchain-ollama
python-dotenv
pydantic>=2.0.0
langchain-community>=0.2.0,<0.4.0
starlette
fastapi
click
uvicorn
httpx
asyncclick